modificar forma de moviemiento de inventario en facturacion

+++ db/migrate/20200305203246_create_secuencia_comprobantes.rb
@@ -2,9 +2,9 @@ class CreateSecuenciaComprobantes < ActiveRecord::Migration[5.2]
   def change
     create_table :secuencia_comprobantes do |t|
       t.references :tipo_factura, foreign_key: true
-      t.integer :secuencia
-      t.integer :desde
-      t.integer :hasta
+      t.bigint :secuencia
+      t.bigint :desde
+      t.bigint :hasta
       t.timestamp :fecha_compra
       t.timestamp :fecha_valida
       t.boolean :estado
